ci: Make bleeding job more robust
Fetch the repository version in first job, reuse it to inject in all
other jobs and use the revision for downloading.

This ensures that all jobs get the exactly same Weblate version. Before
it was possible that different architectures and different steps used
different revisions. This often lead to publishing failure as the image
which should be published was not present in the build cache and build
failed (because we do not apply the qemu build hack to the publish
stage).